find_duplicates_in_dirs: remove inacurate comment.
Expanding the values using d.iteritems() or even d.items() doesn't mean
creating a copy of the values. There wouldn't be a problem with huge
file lists, as the dictionary will just give us a reference to the
existing file list. The problem would of course be if there are many
file lists (many keys). That's already covered in the second portion of
the comment.